feat: Add Apple icon to ExpenseCard component
This commit adds the Apple icon to the ExpenseCard component. The icon is displayed using the node-emoji library and is rendered within a View component. The icon is styled with a text size of 4xl and is centered within the View. This enhancement improves the visual representation of the ExpenseCard component.